move properties to separate module
