+++ extension/pom.xml
@@ -11,6 +11,7 @@
   <packaging>pom</packaging>
 
   <modules>
+    <module>properties</module>
     <module>starter</module>
     <module>starter-rest</module>
     <module>starter-webapp-core</module>
