+++ extension/starter-rest/src/test/java/org/camunda/bpm/extension/spring/boot/rest/CamundaBpmRestConfigurationIT.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.assertEquals;
 
 import org.camunda.bpm.engine.rest.dto.repository.ProcessDefinitionDto;
 import org.camunda.bpm.extension.spring.boot.rest.test.TestRestApplication;
-import org.camunda.bpm.spring.boot.starter.property.CamundaBpmProperties;
+import org.camunda.bpm.spring.boot.properties.CamundaBpmProperties;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
