+++ extension/starter-webapp-core/src/test/java/org/camunda/bpm/spring/boot/starter/webapp/TestApplication.java
@@ -1,6 +1,6 @@
 package org.camunda.bpm.spring.boot.starter.webapp;
 
-import org.camunda.bpm.spring.boot.starter.property.CamundaBpmProperties;
+import org.camunda.bpm.spring.boot.properties.CamundaBpmProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
