+++ extension/starter/src/main/java/org/camunda/bpm/spring/boot/starter/CamundaBpmConfiguration.java
@@ -36,7 +36,7 @@ import org.camunda.bpm.spring.boot.starter.configuration.impl.DefaultProcessEngi
 import org.camunda.bpm.spring.boot.starter.configuration.impl.EnterLicenseKeyConfiguration;
 import org.camunda.bpm.spring.boot.starter.configuration.impl.GenericPropertiesConfiguration;
 import org.camunda.bpm.spring.boot.starter.jdbc.HistoryLevelDeterminator;
-import org.camunda.bpm.spring.boot.starter.property.CamundaBpmProperties;
+import org.camunda.bpm.spring.boot.properties.CamundaBpmProperties;
 import org.camunda.bpm.spring.boot.starter.util.CamundaBpmVersion;
 import org.camunda.bpm.spring.boot.starter.util.CamundaSpringBootUtil;
 import org.springframework.beans.factory.annotation.Qualifier;
