+++ extension/starter/src/main/java/org/camunda/bpm/spring/boot/starter/configuration/impl/DefaultDatasourceConfiguration.java
@@ -4,7 +4,7 @@ import javax.sql.DataSource;
 
 import org.camunda.bpm.engine.spring.SpringProcessEngineConfiguration;
 import org.camunda.bpm.spring.boot.starter.configuration.CamundaDatasourceConfiguration;
-import org.camunda.bpm.spring.boot.starter.property.DatabaseProperty;
+import org.camunda.bpm.spring.boot.properties.DatabaseProperty;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.transaction.PlatformTransactionManager;
