+++ extension/starter/src/test/java/org/camunda/bpm/spring/boot/starter/configuration/impl/DefaultHistoryConfigurationTest.java
@@ -7,7 +7,7 @@ import static org.mockito.Mockito.verify;
 import org.camunda.bpm.engine.impl.history.HistoryLevel;
 import org.camunda.bpm.engine.impl.history.handler.HistoryEventHandler;
 import org.camunda.bpm.engine.spring.SpringProcessEngineConfiguration;
-import org.camunda.bpm.spring.boot.starter.property.CamundaBpmProperties;
+import org.camunda.bpm.spring.boot.properties.CamundaBpmProperties;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
