+++ extension/starter/src/test/java/org/camunda/bpm/spring/boot/starter/jdbc/HistoryLevelDeterminatorJdbcTemplateImplTestApplication.java
@@ -3,7 +3,7 @@ package org.camunda.bpm.spring.boot.starter.jdbc;
 import javax.sql.DataSource;
 
 import org.camunda.bpm.spring.boot.starter.CamundaBpmAutoConfiguration;
-import org.camunda.bpm.spring.boot.starter.property.CamundaBpmProperties;
+import org.camunda.bpm.spring.boot.properties.CamundaBpmProperties;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.jdbc.core.JdbcTemplate;
